requirements.txt
tk
pillow
PyPDF2